# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
...
paths:
...
  /campaigns:
    get:
      description: handles retrieval of campaigns
      operationId: getCampaigns
      parameters:
        - in: query
          name: status
          description: Filter campaigns by status. Multiple status values can be specified by repeating the parameter
          required: false
          schema:
            type: array
            items:
              type: string
              enum: ["scheduled", "running", "paused", "cancelled"]
        - in: query
          name: no_body
          description: When set to true, returns response without body content
          required: false
          schema:
            type: boolean
        - in: query
          name: page
          description: Page number for paginated results.
          required: false
          schema:
            type: integer
        - in: query
          name: per_page
          description: Number of items per page. Use an integer for specific page size or 'all' to retrieve all results
          required: false
          schema:
            oneOf:
              - type: integer
                description: Number of items to return per page
              - type: string
                enum: ["all"]
                description: Return all results without pagination
        - in: query
          name: tags
          description: Filter campaigns by tags. Multiple tags can be specified by repeating the parameter
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: order
          description: Determines the sort order of results. ASC for ascending, DESC for descending order
          required: false
          schema:
            type: string
            enum: ["ASC", "DESC"]
        - in: query
          name: order_by
          description: Specifies the field by which to sort the campaigns. Available options are 'name', 'status', 'created_at', and 'updated_at'
          required: false
          schema:
            type: string
            enum: ["name", "status", "created_at", "updated_at"]
        - in: query
          name: query
          description: SQL query expression to filter campaigns by custom criteria
          required: false
          schema:
            type: string
      tags:
        - Campaigns
      responses:
        "200":
          description: list of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/Campaign"
                      query:
                        type: string
                      total:
                        type: integer
                      per_page:
                        type: integer
                      page:
                        type: integer

    post:
      description: handles campaign creation
      operationId: createCampaign
      tags:
        - Campaigns
      requestBody:
        description: new campaign info
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignRequest"

      responses:
        "200":
          description: new campaign object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CampaignUpdate"
  "/campaigns/{id}":
    get:
      description: handles retrieval of campaigns.
      operationId: getCampaignById
      tags:
        - Campaigns
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the campaign you want to get.
          schema:
            type: integer
        - in: query
          name: no_body
          required: false
          description: boolean flag for response with/without body
          schema:
            type: boolean
      responses:
        "200":
          description: campaign object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Campaign"

    put:
      description: handle updation of campaign
      operationId: updateCampaignById
      tags:
        - Campaigns
      parameters:
        - in: path
          name: id
          required: true
          description: the id value of campaign you want to update
          schema:
            type: integer
      requestBody:
        description: updated campaign fields
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignRequest"
      responses:
        "200":
          description: updated campaign object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CampaignUpdate"

    delete:
      description: deletes specified campaign
      operationId: deleteCampaignById
      tags:
        - Campaigns
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the campaign you want to get.
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  /campaigns/running/stats:
    get:
      description: returns stats of a given set of campaign IDs.
      operationId: getRunningCampaignStats
      tags:
        - Campaigns
      parameters:
        - in: query
          name: campaign_id
          description: Campaign IDs to get stats for.
          required: true
          schema:
            type: number
      responses:
        "200":
          description: list of stats for given set of campaign ids
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CampaignStats"
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        CampaignID:
          type: integer
        views:
          type: integer
        clicks:
          type: integer
        lists:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        started_at:
          type: string
        to_send:
          type: integer
        sent:
          type: integer
        uuid:
          type: string
        type:
          type: string
          enum: [regular, optin]
        name:
          type: string
        subject:
          type: string
        from_email:
          type: string
        body:
          type: string
        send_at:
          type: string
        status:
          type: string
        content_type:
          type: string
          enum: [richtext, html, markdown, plain]
        tags:
          type: array
          items:
            type: string
        template_id:
          type: integer
        messenger:
          type: string
...
    CampaignRequest:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        lists:
          type: array
          items:
            type: number
        from_email:
          type: string
        content_type:
          type: string
        messenger:
          type: string
        type:
          type: string
        tags:
          type: array
          items:
            type: string
        send_later:
          type: boolean
        send_at:
          type: object
          properties:
            headers:
              type: array
              items:
                type: object
            template_id:
              type: number
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        lists:
          type: array
          items:
            type: integer
        from_email:
          type: string
        messenger:
          type: string
        type:
          type: string
        tags:
          type: array
          items:
            type: string
        send_later:
          type: boolean
        send_at:
          type: object
        headers:
          type: array
          items:
            type: object
        template_id:
          type: integer
        content_type:
          type: string
        body:
          type: string
        altbody:
          type: string
        archive:
          type: boolean
        archive_template_id:
          type: integer
        archive_meta:
          type: object
    CampaignStats:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        to_send:
          type: integer
        sent:
          type: integer
        started_at:
          type: string
          format: date
        updated_at:
          type: string
          format: date
        rate:
          type: integer
        net_rate:
          type: integer
...